ai-chain-app/requirements.txt
streamlit
requests
langchain-google-vertexai
langchain-core
google-cloud-aiplatform